Ask me for the fix for Fedora 22 :)
Reverting...
Fix for Fedora 22 is the same as for Ubuntu 15.04.  You forgot to include Centos 7 into the osmajversion section so it was an issue for our Centos 7 machines.  
We'll utilize the site.pp with some case statements to run Service { provider => 'systemd' } for Ubuntu and Fedora since currently only puppet 4.x or 4.1 supports them properly.